Fixes : #28 Header and C files for measurements related function
